requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
SQLAlchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
qrcode[pil]==7.4.2